refactor: simplify session storage
